refactor(webhook): extract app-lifecycle D1 helpers into mutations-app.ts
Split mutations.ts (302 SLOC, was file-length-exempt) into two cohesive
modules under the 300-SLOC cap:
- mutations.ts (160): issue/edge/PR/milestone "corpus" write helpers
- mutations-app.ts (142): GitHub App install-lifecycle writes (tenant,
  repo access, repo registry + cascade rename, cache invalidation,
  session/install-token cleanup)

Pure move, no logic change. Importers updated to direct sibling imports
(handlers-app.ts, mutations.test.ts); bumpDataVersion stays in mutations.ts.
Removes the mutations.ts file-length exemption.
